moved last action to api
