+++ frontend/lost/src/api/worker.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query'
-import { httpClient } from '../actions/http-client'
+import { httpClient } from './http_client'
 
 export const useWorkers = () => {
   return useQuery({
