+++ frontend/lost/src/containers/Authentication/OidcCallbackHandler.tsx
@@ -3,7 +3,7 @@ import axios from 'axios'
 import { jwtDecode } from 'jwt-decode'
 import { useEffect, useRef, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { exchangeOidcCode } from '../../api/auth_api'
+import { exchangeOidcCode } from '../../api/auth/auth'
 
 /**
  * OidcCallbackHandler
