+++ frontend/lost/src/coreui_containers/TheFooter.tsx
@@ -2,7 +2,7 @@ import { CFooter } from '@coreui/react'
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useInterval } from 'react-use'
-import { checkExpireDate, getRemainingLoggedInSeconds } from '../actions/auth'
+import { checkExpireDate, getRemainingLoggedInSeconds } from '../api/auth/auth_utils'
 import useInactive from '../hooks/useInactive'
 import { useLostConfig } from '../hooks/useLostConfig'
 import LogoutModal from '../containers/Authentication/LogoutModal'
